feat: Refactor utility functions and remove unused code from user models
